split detect
